requirements.txt
streamlit
transformers
torch==2.1.2+cpu
torchvision==0.16.2+cpu
Pillow
numpy
requests